requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
requests
httpx
aiohttp
py-healthcheck